chore: rename files to match conventional pattern
